# rpi5_snapshot/openapi.yaml
openapi: 3.0.3
info:
  title: Game Stats API
  version: 1.0.0
  description: API kontrakty dla modułów 1v1 oraz multiplayer.
servers:
  - url: /api/v1
components:
  schemas:
    Uuid:
      type: string
      format: uuid
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    Date:
      type: string
      format: date
      example: "2026-01-30"
    DateTime:
      type: string
      format: date-time
      example: "2026-01-30T12:00:00.000Z"
    ErrorDetailsItem:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorPayload:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, NOT_FOUND, CONFLICT, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetailsItem'
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorPayload'
...
    TrainsCounts:
      type: object
      description: "Klucze '1'..'9' z wartościami integer >= 0"
      required: ['1','2','3','4','5','6','7','8','9']
      additionalProperties: false
      properties:
        '1': { type: integer, minimum: 0 }
        '2': { type: integer, minimum: 0 }
        '3': { type: integer, minimum: 0 }
        '4': { type: integer, minimum: 0 }
        '5': { type: integer, minimum: 0 }
        '6': { type: integer, minimum: 0 }
        '7': { type: integer, minimum: 0 }
        '8': { type: integer, minimum: 0 }
        '9': { type: integer, minimum: 0 }
    TicketToRideVariant:
      type: object
      required: [id, code, name]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
          example: "europe"
        name:
          type: string
          example: "Europa"
...
    TicketToRideMatchPlayer:
      type: object
      required:
        [id, player, ticketsPoints, bonusPoints, trainsCounts, trainsPoints, totalPoints, place]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        player:
          type: object
          required: [id, name]
          properties:
            id:
              $ref: '#/components/schemas/Uuid'
            name:
              type: string
        ticketsPoints:
          type: integer
        bonusPoints:
          type: integer
        trainsCounts:
          $ref: '#/components/schemas/TrainsCounts'
        trainsPoints:
          type: integer
        totalPoints:
          type: integer
        place:
          type: integer
    TicketToRideMatch:
      type: object
      required: [id, playedOn, variant, players]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        playedOn:
          $ref: '#/components/schemas/Date'
        variant:
          $ref: '#/components/schemas/TicketToRideVariant'
        notes:
          type: string
          nullable: true
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        players:
          type: array
          items:
            $ref: '#/components/schemas/TicketToRideMatchPlayer'
    TicketToRideMatchListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TicketToRideMatch'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    TicketToRideMatchCreate:
      type: object
      required: [playedOn, variantId, players]
      properties:
        playedOn:
          $ref: '#/components/schemas/Date'
        variantId:
          $ref: '#/components/schemas/Uuid'
        notes:
          type: string
          maxLength: 2000
          nullable: true
        players:
          type: array
          minItems: 2
          maxItems: 5
          items:
            type: object
            required: [playerId, ticketsPoints, bonusPoints, trainsCounts]
            properties:
              playerId:
                $ref: '#/components/schemas/Uuid'
              ticketsPoints:
                type: integer
              bonusPoints:
                type: integer
                minimum: 0
              trainsCounts:
                $ref: '#/components/schemas/TrainsCounts'
    TicketToRideMatchPatch:
      type: object
      properties:
        playedOn:
          $ref: '#/components/schemas/Date'
        variantId:
          $ref: '#/components/schemas/Uuid'
        notes:
          type: string
          maxLength: 2000
          nullable: true
        players:
          type: array
          minItems: 2
          maxItems: 5
          items:
            type: object
            required: [playerId, ticketsPoints, bonusPoints, trainsCounts]
            properties:
              playerId:
                $ref: '#/components/schemas/Uuid'
              ticketsPoints:
                type: integer
              bonusPoints:
                type: integer
                minimum: 0
              trainsCounts:
                $ref: '#/components/schemas/TrainsCounts'
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /ticket-to-ride/matches:
    get:
      summary: List Ticket to Ride matches (legacy endpoint)
      parameters:
        - in: query
          name: dateFrom
          schema:
            $ref: '#/components/schemas/Date'
        - in: query
          name: dateTo
          schema:
            $ref: '#/components/schemas/Date'
        - in: query
          name: variantId
          schema:
            $ref: '#/components/schemas/Uuid'
        - in: query
          name: playerId
          schema:
            $ref: '#/components/schemas/Uuid'
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketToRideMatchListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create Ticket to Ride match (legacy endpoint)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketToRideMatchCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketToRideMatch'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /ticket-to-ride/matches/{id}:
    get:
      summary: Get Ticket to Ride match by id (legacy endpoint)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketToRideMatch'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      summary: Update Ticket to Ride match (legacy endpoint)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketToRideMatchPatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketToRideMatch'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete Ticket to Ride match (legacy endpoint)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '204':
          description: No content
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...